src/requirements.txt
aiofiles==23.2.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
packaging==24.1
pluggy==1.5.0
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
pypika-tortoise==0.1.6
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
text-unidecode==1.3
tomli==2.0.1
tortoise-orm==0.21.3
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
